commons/limitup-ib-common-api/src/main/java/com/limituptrading/ib/IBClient.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.limituptrading.ib;

/**
 *
 * @author RobTerpilowski
 */
public class IBClient {
 
    protected String host;
    protected int port;
    protected int clientId;

    public IBClient(String host, int port, int clientId) {
        this.host = host;
        this.port = port;
        this.clientId = clientId;
    }

    public String getHost() {
        return host;
    }

    public int getPort() {
        return port;
    }

    public int getClientId() {
        return clientId;
    }
    
    
    
}
